Create libqubes-pure: a library of Qubes-specific code that does no I/O
Currently this only includes functions for sanitizing strings and paths,
but future versions will also support validating qube names, qrexec
service names and arguments, and possibly more.  The header file also
includes types for string buffers and extensive documentation.
