+++ debian/libqubes-rpc-filecopy-dev.install
@@ -1,2 +1,4 @@
 usr/include/libqubes-rpc-filecopy.h
+usr/include/qubes/pure.h
 usr/lib/libqubes-rpc-filecopy.so
+usr/lib/libqubes-pure.so
