+++ debian/libqubes-rpc-filecopy2.install
@@ -1,1 +1,2 @@
 usr/lib/libqubes-rpc-filecopy.so.2*
+usr/lib/libqubes-pure.so.0*
