+++ debian/libqubes-rpc-filecopy2.shlibs
@@ -1,1 +1,2 @@
 libqubes-rpc-filecopy 2 libqubes-rpc-filecopy2 (>= 4.1.7)
+libqubes-pure 0 libqubes-rpc-filecopy2 (>= 4.2.0)
